script temp dir can be specified in cli

+++ src/include/osiris/internal/temp/donetemp.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 _osiris_internal_temp_donetemp_main() {
-	if [ -d "${SCRIPT_TEMP_DIR}" ]; then
+	if [ -z "${KEEP_SCRIPT_TEMP_DIR}" ] && [ -d "${SCRIPT_TEMP_DIR}" ]; then
 		rm -rf --preserve-root "${SCRIPT_TEMP_DIR}"
 	fi
 }
